+++ gpt_disk_io/tests/test_block.rs
@@ -81,3 +81,19 @@ fn test_block_size() {
     assert_eq!(BlockSize::default().to_u32(), 512);
     assert_eq!(BlockSize::default().to_u64(), 512);
 }
+
+#[test]
+fn test_block_size_is_multiple() {
+    assert!(BlockSize::BS_512.is_multiple_of_block_size(0));
+    assert!(BlockSize::BS_512.is_multiple_of_block_size(512));
+    assert!(BlockSize::BS_512.is_multiple_of_block_size(1024));
+
+    assert!(!BlockSize::BS_512.is_multiple_of_block_size(1023));
+    assert!(!BlockSize::BS_512.is_multiple_of_block_size(1025));
+}
+
+#[test]
+#[should_panic]
+fn test_block_size_is_multiple_panic() {
+    let _ = BlockSize::BS_512.is_multiple_of_block_size(u128::MAX);
+}